improve error throwing way,remove console.log,add lastViewat or modifiedat

+++ app/api/file/route.js
@@ -6,7 +6,7 @@ import {
   getFolder,
 } from "@/lib/database";
 import { fileAllowExtension, type2extensionDictionary } from "@/lib/constant";
-import { NextResponse } from "next/server";
+
 export const GET = async (request, { params }) => {
   const url = new URL(request.url);
   const searchParams = new URLSearchParams(url.search);
@@ -21,17 +21,18 @@ export const GET = async (request, { params }) => {
       (item) => item.name === name && item.extension === extension
     )?._id;
     const file = await File.findById(fileId);
-    await File.findByIdAndUpdate(file._id, { lastViewAt: new Date() });
+    await File.findByIdAndUpdate(file._id, {
+      lastViewAt: new Date(),
+    });
     await Folder.findByIdAndUpdate(parentFolder._id, {
       lastViewAt: new Date(),
     });
-    //dont send complete file,only part of fields
     return new Response(JSON.stringify({ content: file.base64String }), {
       status: 200,
     });
   } catch (error) {
     return new Response(JSON.stringify({ message: error.message }), {
-      status: 500,
+      status: error.code || 500,
     });
   }
 };
@@ -40,10 +41,8 @@ export const POST = async (request, { params }) => {
   const req = await request.json();
   try {
     if (!fileAllowExtension.includes(req.extension)) {
-      console.log("file type not allow");
       throw new Error("file type not allowed");
     }
-
     await connectToDB();
     const pathList = req.path?.split("/");
     await createRootFolderIfNotExist(pathList[0]);
@@ -53,7 +52,6 @@ export const POST = async (request, { params }) => {
         (file) => file.name === req.name && file.extension === req.extension
       )
     ) {
-      console.log("detect duplicated file");
       throw new Error("file already exist");
     }
     const file = new File({
@@ -71,27 +69,21 @@ export const POST = async (request, { params }) => {
       extension: new_file.extension,
       path: new_file.path,
     });
-    const new_parent_folder = await Folder.findByIdAndUpdate(
-      parentFolder._id,
-      parentFolder,
-      { new: true }
-    );
+    parentFolder.lastViewAt = new Date();
+    await Folder.findByIdAndUpdate(parentFolder._id, parentFolder, {
+      new: true,
+    });
     return new Response(JSON.stringify({ data: "success" }), { status: 200 });
   } catch (error) {
-    return new Response(
-      JSON.stringify("Failed to fetch prompts created by user"),
-      {
-        status: 500,
-      }
-    );
+    return new Response(JSON.stringify({ message: error.message }), {
+      status: error.code || 500,
+    });
   }
 };
 
 export const PUT = async (request) => {
   try {
-    console.log("receive file update request");
     const req = await request.json();
-    console.log(req);
     const extension = type2extensionDictionary[req.type];
     await connectToDB();
     const pathList = req.path?.split("/");
@@ -99,23 +91,26 @@ export const PUT = async (request) => {
     const file = parentFolder.fileList.find(
       (file) => file.name === req.name && file.extension === extension
     );
-    console.log("get file", file);
+    if (!file) {
+      throw new Error("file not exist");
+    }
     if (req.comments) {
       await File.findByIdAndUpdate(file._id, {
         comments: req.comments,
         modifiedAt: new Date(),
+        lastViewAt: new Date(),
       });
     } else if (req.base64String) {
       await File.findByIdAndUpdate(file._id, {
         base64String: req.base64String,
         modifiedAt: new Date(),
+        lastViewAt: new Date(),
       });
     }
-
     return new Response(JSON.stringify({ data: "success" }), { status: 200 });
   } catch (error) {
-    return new Response("Failed to update", {
-      status: 500,
+    return new Response(JSON.stringify({ message: error.message }), {
+      status: error.code || 500,
     });
   }
 };
@@ -123,33 +118,27 @@ export const PUT = async (request) => {
 export const DELETE = async (request) => {
   try {
     const req = await request.json();
-    console.log("receive file delete request", req);
     await connectToDB();
     const pathList = req.path?.split("/");
     const parentFolder = await getFolder(pathList, 1, pathList.length);
-    console.log("get parent folder", parentFolder);
     const fileObj = parentFolder?.fileList?.find(
       (item) => item.name === req.name
     );
-
-    console.log("get file", fileObj);
+    if (!fileObj) {
+      throw new Error("file not existed");
+    }
     await File.deleteOne({ _id: fileObj._id });
-    console.log("after delete file");
-    const newfolder = await Folder.findByIdAndUpdate(
-      parentFolder._id,
-      {
-        fileList: parentFolder.fileList.filter(
-          (file) => file.name !== fileObj.name
-        ),
-        modifiedAt: new Date(),
-      },
-      { new: true }
-    );
-    console.log("after delte from folder", newfolder);
+    await Folder.findByIdAndUpdate(parentFolder._id, {
+      fileList: parentFolder.fileList.filter(
+        (file) => file.name !== fileObj.name
+      ),
+      modifiedAt: new Date(),
+      lastViewAt: new Date(),
+    });
     return new Response(JSON.stringify({ data: "success" }), { status: 200 });
   } catch (error) {
-    return new Response("Error during delete", {
-      status: 500,
+    return new Response(JSON.stringify({ message: error.message }), {
+      status: error.code || 500,
     });
   }
 };
